Split state classes into different modules
