+++ poliastro/twobody/__init__.py
@@ -1,3 +1,3 @@
 # coding: utf-8
 
-from poliastro.twobody.state import State
+from poliastro.twobody.core import State
